// src/features/agency-register/api/api.d.ts
export interface components {
    schemas: {
        /** @description Binnenlands of buitenlands adres */
        Adres: {
            binnenlandsAdres?: components["schemas"]["BinnenlandsAdres"];
            buitenlandsAdres?: components["schemas"]["BuitenlandsAdres"];
        };
        /**
         * @description Bezoekadres of postadres
         * @enum {string}
         */
        AdresType: "bezoekadres" | "postadres";
        BinnenlandsAdres: {
            type?: components["schemas"]["AdresType"];
            straatnaam?: string;
            /** Format: int32 */
            huisnummer?: number;
            huisletter?: string;
            /**
             * Format: int32
             * @description Postbusnummer wordt alleen getoond indien het een postadres betreft
             */
            postbusnummer?: number;
            postcode?: string;
            plaats?: string;
        };
        BuitenlandsAdres: {
            type?: components["schemas"]["AdresType"];
            /**
             * @description Het straat/huisnummer is een combinatie van de straat en huisnummer
             * @example 53 Rue de Tilsitt
             */
            straatHuisnummer?: string;
            /**
             * @description De postcode/woonplaats is de combinatie van een eventuele postcode en woonplaats
             * @example X-13501 Parijs
             */
            postcodeWoonplaats?: string;
            /**
             * @description De naam van het land waar het adres bevindt
             * @example Frankrijk
             */
            land?: string;
        };
        Links: {
            [key: string]: components["schemas"]["Link"];
        };
        Resultaat: {
            /**
             * Format: int32
             * @description Geeft aan op welke pagina je bent. Start vanaf pagina 1
             */
            pagina?: number;
            /**
             * Format: int32
             * @description Geeft het aantal zoek resultaten per pagina weer
             */
            resultatenPerPagina?: number;
            /**
             * Format: int32
             * @description Totaal aantal zoekresultaten gevonden. De API Zoeken toont max. 1000 resultaten.
             */
            totaal?: number;
            /** @description Link naar de vorige pagina indien beschikbaar */
            vorige?: string;
            /** @description Link naar de volgende pagina indien beschikbaar */
            volgende?: string;
            resultaten?: components["schemas"]["ResultaatItem"][];
            _links?: components["schemas"]["Links"];
        };
        ResultaatItem: {
            /** @description Nederlands Kamer van Koophandel nummer dat bestaat uit 8 cijfers */
            kvkNummer?: string;
            /** @description Rechtspersonen Samenwerkingsverbanden Informatie Nummer dat bestaat uit 9 cijfers */
            rsin?: string;
            /** @description Vestigingsnummer dat bestaat uit 12 cijfers */
            vestigingsnummer?: string;
            /** @description De naam waaronder een vestiging of rechtspersoon handelt */
            naam?: string;
            adres?: components["schemas"]["Adres"];
            /** @description hoofdvestiging/nevenvestiging/rechtspersoon */
            type?: string;
            /** @description Indicatie of inschrijving actief is */
            actief?: string;
            /** @description Bevat de vervallen naam waaronder een vestiging of rechtspersoon heeft gehandeld */
            vervallenNaam?: string;
            _links?: components["schemas"]["Links"];
        };
        Link: {
            href?: string;
            title?: string;
        };
        Error: {
            fout?: components["schemas"]["Fout"][];
        };
        Fout: {
            /** @description Foutcode */
            code?: string;
            /** @description Omschrijving van de foutmelding */
            omschrijving?: string;
        };
    };
// ...
}